chore: remove payment logic out of send screen
